WWCS/dashboard/service/get_open_meteo/requirements.txt
ipython
mysql-connector-python
netCDF4
numpy
openmeteo-requests
pandas
python-dotenv
pyyaml
requests-cache
retry-requests
xarray
